perbaikan bug renstra

+++ public/class-wp-eval-sakip-public.php
@@ -734,6 +734,35 @@ class Wp_Eval_Sakip_Public extends Wp_Eval_Sakip_Verify_Dokumen
 		    return $hak_akses_user;
 	  	}
 
+	  	public function hak_akses_upload_dokumen_renstra($nama_dokumen,)
+	 	{    
+		    global $wpdb;
+		    $current_user = wp_get_current_user();
+		    $user_roles = $current_user->roles;
+		    $is_administrator = in_array('administrator', $user_roles);
+
+		    $admin_role_pemda = array(
+		      'admin_bappeda',
+		      'admin_ortala'
+		    );
+
+		      $this_jenis_role = (in_array($user_roles[0], $admin_role_pemda)) ? 1 : 2 ;
+
+		    $cek_settingan_menu = $wpdb->get_var(
+		      $wpdb->prepare(
+		      "SELECT 
+		        jenis_role
+		      FROM esakip_menu_dokumen 
+		      WHERE nama_dokumen=%s
+		      AND user_role='perangkat_daerah' 
+		      AND active = 1
+		    ", $nama_dokumen)
+		    );
+
+		    $hak_akses_user = ($cek_settingan_menu == $this_jenis_role || $cek_settingan_menu == 3 || $is_administrator) ? true : false;
+		    return $hak_akses_user;
+	  	}
+
 	public function dokumen_detail_dpa_pemda($atts)
 	{
 		// untuk disable render shortcode di halaman edit page/post
@@ -6281,7 +6310,7 @@ class Wp_Eval_Sakip_Public extends Wp_Eval_Sakip_Verify_Dokumen
 						// 	$btn .= '<button class="btn btn-sm btn-success" onclick="verifikasi_dokumen(\'' . $vv['id'] . '\'); return false;" href="#" title="Verifikasi Dokumen"><span class="dashicons dashicons-yes"></span></button>';
 						// }
 						// if (!$this->is_admin_panrb() && !$this->is_admin_user()) {
-						if (!$this->is_admin_panrb()) {
+						if (!$this->is_admin_panrb() && $this->hak_akses_upload_dokumen_renstra('RENSTRA')) {
 							$btn .= '<button class="btn btn-sm btn-warning" onclick="edit_dokumen_renstra(\'' . $vv['id'] . '\'); return false;" href="#" title="Edit Dokumen"><span class="dashicons dashicons-edit"></span></button>';
 							$btn .= '<button class="btn btn-sm btn-danger" onclick="hapus_dokumen_renstra(\'' . $vv['id'] . '\'); return false;" href="#" title="Hapus Dokumen"><span class="dashicons dashicons-trash"></span></button>';
 						}
@@ -11438,7 +11467,7 @@ class Wp_Eval_Sakip_Public extends Wp_Eval_Sakip_Verify_Dokumen
 
 						$btn = '<div class="btn-action-group">';
 						$btn .= '<button class="btn btn-info" onclick="lihatDokumen(\'' . $vv['dokumen'] . '\'); return false;" href="#" title="Lihat Dokumen"><span class="dashicons dashicons-visibility"></span></button>';
-						if (!$this->is_admin_panrb() && !$this->is_admin_user()) {
+						if (!$this->is_admin_panrb() && $this->hak_akses_upload_dokumen_renstra('RENSTRA')) {							
 							$btn .= "<button class='btn btn-success' onclick='set_tahun_dokumen(" . $vv['id'] . "); return false;' title='Set Tahun Dokumen'><span class='dashicons dashicons-insert'></span></button>";
 
 							$btn .= '<button class="btn btn-danger" onclick="hapus_tahun_dokumen_renstra(\'' . $vv['id'] . '\'); return false;" href="#" title="Hapus Dokumen"><span class="dashicons dashicons-trash"></span></button>';
